UQMAnimationTool: Added playback functionality, so that you can play animations that you've defined. See the 'playback' tab on the right. Currently, there may be some Z-Indexing problems which may mean that the reproduction of an animation from UQMAnimationTool may not be completely accurate when compared with UQM's output, especially when frames are overlapping each other.
UQMAnimationTool sets the Z-Index for all images based on their position in the frame definition list (frames defined at the top of the .ani file are painted 'below' frames defined at the bottom of the .ani file). On every tick of the update loop, the animation system will loop through all defined frames and check if they should be visible or not based on all animations currently running. UQM, on the other hand, seems to have a system where frames are painted on the screen when needed, on top of the active screen buffer. Will have to fix that in a future update, or will have to introduce some z-index overriding feature that allow artists to properly set the Z-indexes of frames manually (SMIL forces setting of z-indexes, but otherwise simulates the way UQM works quite well if all z-indexes of regions are the same).